Extend package to include Places 1.2
The latest master only included support for SimpleGeo Places 1.0. This
commit includes support for Places 1.2. This was done by extending the
SimpleGeo prototype to SimpleGeo.Places and SimpleGeo.Places12,
respectively. The README has been updated to include information on
how to use both versions of Places.

You must still use Places 1.0 if you intend to use getNearbyPlacesByLatLng,
getNearbyPlacesByAddress, and getNearbyPlacesByIp. You should use
Places 1.2 for getNearbyPlacesFromBBox and getNearbyPlacesText. The only
overlap that exists is in the getNearbyPlaces function to stay as DRY
as possible.

Any suggestions on how to improve this commit would be seriously
appreciated.

Signed-off-by: Josh Smith <joshdotsmith@gmail.com>